Migrated from AirTableClient to MySQLClient and Web API backend. - All the AirTable requests still need to be migrated. - The Web API currently only has Authorization set up. - Changed a few small things related to matchmaking.

+++ SquidDraftLeague/SquidDraftLeague.Draft/Matchmaking/Matchmaker.cs
@@ -34,7 +34,7 @@ namespace SquidDraftLeague.Draft.Matchmaking
             new Set(5)
         }.AsReadOnly();
 
-        public static SdlClass GetClass(double powerLevel)
+        public static SdlClass GetClass(decimal powerLevel)
         {
             if (powerLevel >= 2200)
             {
